Fix mysql container
The mysql container had some startup problems. First of all, the startup
script was not included inside of the setup.py file. This should now be
fixed.

Moreover, the container had some permission problems. This was caused by
the custom *custom.cnf* file that is used during the mysqld startup. I'm
not totally sure which option was causing the issue but I just
substituted the complete configuration file with a different version to
fix it.