Print ssh container IP
Since the main purpose of the ssh container is port forwarding, it is
often required to know the IP address of this container. When forwarding
a remote port, it will be accessible only on the containers IP address
and not on the host system. Since I found myself using *docker inspect*
way to many times, I guess that printing the IP address on startup could
be helpful.