Fix neo4j password bug
Fixed a bug inside of the neo4j startup script. Previously, the
set-initial-password function was used for password resets. However,
this function only works on the intial startup. For any further
startup, the password needs to be changed inside the database.

Changing passwords inside of the database for further neo4j runs was
now implemented. This commit fixes issue #8.