Add option to allow root login on ssh
Some tools do not work correctly if software with default ports (like
SMB) is bound to a different port (e.g. mapping 4445 -> 445). In these
cases you actually want a mapping (445 -> 445), which is only possible
with root privileges.

Enabling root login per default on the container is of course
questionable. Therefore, a script ``toggle-root.sh`` was added. To
enable root login on the ssh container, one can now do the following:

1. Mirror the container ``car mirror ssh``
2. Toggle root login ``cd car && bash toggle-root.sh``
3. Run the mirrored container ``car run .``