Remove car-network for all containers
The car network was introduced 2 commits before this one to put all
containers launched by car into a shared network. The idea was to reduce
the amount of bridge interfaces created by docker when running multiple
car containers at once.

However, after thinking about this, I decided that this is not a good
idea. Container networks are seperated by default to isolate them.
Putting unrelated containers that do not need to communicate with each
other into a shared network breaks this isolation and could have
security implications. Of course, with the purpose of car it is very
unlikely that this causes issues, but following bad practices should not
be the default configuration.

To prevent many dangling docker-bridges, just make sure to run ``car rm
<container>`` after finishing with a container or run ``car rm all`` to
cleanup all.