Update ajp container
Updated the ajp container. The container uses now a fixed version of
alpine and installs apache2 manually. This way, the overall container
size could be reduced from 200MB to about 10MB. The only downside is
that mod_jk.so now need to be compiled manually. However, a script was
added to do this automatically and also a precompiled version was
included.