Change folder delete method
During the 'car clean' operation, 'shutil.rmtree' was used so far. This
one has problems when files of different users are located in a user
controlled folder (resulting in a permission denied error). With 'rm -r'
no such error occurs, which is more handy for 'car clean'. Therefore,
the call was changed to 'subprocess.call(['rm', '-r', ...)'.