Update tftp container
Updated the logging mechanisms of the tftp container. Finally figured
out to get it done correctly. tftpd expects ``/dev/log`` to be a unix
socket as it is created by syslogd. Since busybox is already installed
on the container, we can use busybox minimal syslogd service to get the
logs.